requirements.txt
asgiref==3.7.2
Django==4.2.1
django-globals==0.3.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
loguru==0.7.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.2